Refactor serialize into separate file
