+++ osgar/drivers/test_bus.py
@@ -116,10 +116,4 @@ class BusHandlerTest(unittest.TestCase):
         bus.publish('new_channel', b'some data')
         self.assertEqual(bus.listen(), (timedelta(microseconds=11), 'raw', (3, 4, 5)))
 
-    def test_serialization(self):
-            self.assertEqual(serialize(b'\x01\x02'), b'\x01\x02')
-            position = (51749517, 180462688)
-            self.assertEqual(serialize(position), b'(51749517, 180462688)')
-            self.assertEqual(serialize((123.4, 'Hi')), b"(123.4, 'Hi')")
-
 # vim: expandtab sw=4 ts=4
